6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int i;
 long int li;
 long long int lli;
 float f;
-printf("Size of char: %ld byte\n",sizeof(c));
-printf("Size of int: %ld bytes\n", sizeof(i));
-printf("Size of long int: %ld bytes\n", sizeof(li));
-printf("Size of long long int: %ld bytes\n", sizeof(lli));
-printf("Size of float: %ld bytes\n", sizeof(f));
+printf("Size of char: %ld byte\n", (unsigned long)sizeof(c));
+printf("Size of int: %ld bytes\n", (unsigned long) sizeof(i));
+printf("Size of long int: %ld bytes\n", (unsigned long) sizeof(li));
+printf("Size of long long int: %ld bytes\n", (unsigned long) sizeof(lli));
+printf("Size of float: %ld bytes\n", (unsigned long) sizeof(f));
 	return (0);
 }
